Move JiraConnection to a separate file
For clarity, move the class JiraConnection to its own module.
This way, it is more clear the purpose of the module jira_issue.py, just
to create new tickets.
